requirements_complete.txt
numpy>=1.21.0
pyyaml>=6.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
openpyxl>=3.0.0
flask>=2.0.0
